move spack.llnl.url into spack.util.url
